Move FuzzerSetRand to separate file.
This fixes Google's OSS-Fuzz builds.
